Introduce ResourceSelector interface to be used by diferent operators
