+++ pkg/prometheus/server/operator.go
@@ -24,15 +24,11 @@ import (
 	"strings"
 	"time"
 
-	"github.com/asaskevich/govalidator"
-	"github.com/blang/semver/v4"
 	"github.com/go-kit/log"
 	"github.com/go-kit/log/level"
 	"github.com/mitchellh/hashstructure"
 	"github.com/pkg/errors"
 	"github.com/prometheus/client_golang/prometheus"
-	"github.com/prometheus/common/model"
-	"github.com/prometheus/prometheus/model/relabel"
 	appsv1 "k8s.io/api/apps/v1"
 	v1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
@@ -1606,17 +1602,18 @@ func (c *Operator) createOrUpdateConfigurationSecret(ctx context.Context, p *mon
 		return nil
 	}
 
-	smons, err := c.selectServiceMonitors(ctx, p, store)
+	resourceSelector := prompkg.NewResourceSelector(c.logger, p, store, c.pmonInfs, c.smonInfs, c.probeInfs, c.nsMonInf, c.metrics)
+	smons, err := resourceSelector.SelectServiceMonitors(ctx)
 	if err != nil {
 		return errors.Wrap(err, "selecting ServiceMonitors failed")
 	}
 
-	pmons, err := c.selectPodMonitors(ctx, p, store)
+	pmons, err := resourceSelector.SelectPodMonitors(ctx)
 	if err != nil {
 		return errors.Wrap(err, "selecting PodMonitors failed")
 	}
 
-	bmons, err := c.selectProbes(ctx, p, store)
+	bmons, err := resourceSelector.SelectProbes(ctx)
 	if err != nil {
 		return errors.Wrap(err, "selecting Probes failed")
 	}
@@ -1811,420 +1808,6 @@ func (c *Operator) createOrUpdateWebConfigSecret(ctx context.Context, p *monitor
 	return nil
 }
 
-func (c *Operator) selectServiceMonitors(ctx context.Context, p *monitoringv1.Prometheus, store *assets.Store) (map[string]*monitoringv1.ServiceMonitor, error) {
-	namespaces := []string{}
-	// Selected object might overlap, deduplicate them by `<namespace>/<name>`.
-	serviceMonitors := make(map[string]*monitoringv1.ServiceMonitor)
-
-	servMonSelector, err := metav1.LabelSelectorAsSelector(p.Spec.ServiceMonitorSelector)
-	if err != nil {
-		return nil, err
-	}
-
-	// If 'ServiceMonitorNamespaceSelector' is nil only check own namespace.
-	if p.Spec.ServiceMonitorNamespaceSelector == nil {
-		namespaces = append(namespaces, p.Namespace)
-	} else {
-		servMonNSSelector, err := metav1.LabelSelectorAsSelector(p.Spec.ServiceMonitorNamespaceSelector)
-		if err != nil {
-			return nil, err
-		}
-
-		namespaces, err = c.listMatchingNamespaces(servMonNSSelector)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	level.Debug(c.logger).Log("msg", "filtering namespaces to select ServiceMonitors from", "namespaces", strings.Join(namespaces, ","), "namespace", p.Namespace, "prometheus", p.Name)
-
-	for _, ns := range namespaces {
-		err := c.smonInfs.ListAllByNamespace(ns, servMonSelector, func(obj interface{}) {
-			k, ok := c.accessor.MetaNamespaceKey(obj)
-			if ok {
-				svcMon := obj.(*monitoringv1.ServiceMonitor).DeepCopy()
-				if err := k8sutil.AddTypeInformationToObject(svcMon); err != nil {
-					level.Error(c.logger).Log("msg", "failed to set ServiceMonitor type information", "namespace", ns, "err", err)
-					return
-				}
-				serviceMonitors[k] = svcMon
-			}
-		})
-		if err != nil {
-			return nil, errors.Wrapf(err, "failed to list service monitors in namespace %s", ns)
-		}
-	}
-
-	var rejected int
-	res := make(map[string]*monitoringv1.ServiceMonitor, len(serviceMonitors))
-	for namespaceAndName, sm := range serviceMonitors {
-		var err error
-
-		for i, endpoint := range sm.Spec.Endpoints {
-			// If denied by Prometheus spec, filter out all service monitors that access
-			// the file system.
-			if p.Spec.ArbitraryFSAccessThroughSMs.Deny {
-				if err = testForArbitraryFSAccess(endpoint); err != nil {
-					break
-				}
-			}
-
-			smKey := fmt.Sprintf("serviceMonitor/%s/%s/%d", sm.GetNamespace(), sm.GetName(), i)
-
-			if err = store.AddBearerToken(ctx, sm.GetNamespace(), endpoint.BearerTokenSecret, smKey); err != nil {
-				break
-			}
-
-			if err = store.AddBasicAuth(ctx, sm.GetNamespace(), endpoint.BasicAuth, smKey); err != nil {
-				break
-			}
-
-			if endpoint.TLSConfig != nil {
-				if err = store.AddTLSConfig(ctx, sm.GetNamespace(), endpoint.TLSConfig); err != nil {
-					break
-				}
-			}
-
-			if err = store.AddOAuth2(ctx, sm.GetNamespace(), endpoint.OAuth2, smKey); err != nil {
-				break
-			}
-
-			smAuthKey := fmt.Sprintf("serviceMonitor/auth/%s/%s/%d", sm.GetNamespace(), sm.GetName(), i)
-			if err = store.AddSafeAuthorizationCredentials(ctx, sm.GetNamespace(), endpoint.Authorization, smAuthKey); err != nil {
-				break
-			}
-
-			if err = validateScrapeIntervalAndTimeout(p, endpoint.Interval, endpoint.ScrapeTimeout); err != nil {
-				break
-			}
-
-			for _, rl := range endpoint.RelabelConfigs {
-				if rl.Action != "" {
-					if err = validateRelabelConfig(*p, *rl); err != nil {
-						break
-					}
-				}
-			}
-
-			for _, rl := range endpoint.MetricRelabelConfigs {
-				if rl.Action != "" {
-					if err = validateRelabelConfig(*p, *rl); err != nil {
-						break
-					}
-				}
-			}
-		}
-
-		if err != nil {
-			rejected++
-			level.Warn(c.logger).Log(
-				"msg", "skipping servicemonitor",
-				"error", err.Error(),
-				"servicemonitor", namespaceAndName,
-				"namespace", p.Namespace,
-				"prometheus", p.Name,
-			)
-			continue
-		}
-
-		res[namespaceAndName] = sm
-	}
-
-	smKeys := []string{}
-	for k := range res {
-		smKeys = append(smKeys, k)
-	}
-	level.Debug(c.logger).Log("msg", "selected ServiceMonitors", "servicemonitors", strings.Join(smKeys, ","), "namespace", p.Namespace, "prometheus", p.Name)
-
-	if pKey, ok := c.accessor.MetaNamespaceKey(p); ok {
-		c.metrics.SetSelectedResources(pKey, monitoringv1.ServiceMonitorsKind, len(res))
-		c.metrics.SetRejectedResources(pKey, monitoringv1.ServiceMonitorsKind, rejected)
-	}
-
-	return res, nil
-}
-
-func (c *Operator) selectPodMonitors(ctx context.Context, p *monitoringv1.Prometheus, store *assets.Store) (map[string]*monitoringv1.PodMonitor, error) {
-	namespaces := []string{}
-	// Selected object might overlap, deduplicate them by `<namespace>/<name>`.
-	podMonitors := make(map[string]*monitoringv1.PodMonitor)
-
-	podMonSelector, err := metav1.LabelSelectorAsSelector(p.Spec.PodMonitorSelector)
-	if err != nil {
-		return nil, err
-	}
-
-	// If 'PodMonitorNamespaceSelector' is nil only check own namespace.
-	if p.Spec.PodMonitorNamespaceSelector == nil {
-		namespaces = append(namespaces, p.Namespace)
-	} else {
-		podMonNSSelector, err := metav1.LabelSelectorAsSelector(p.Spec.PodMonitorNamespaceSelector)
-		if err != nil {
-			return nil, err
-		}
-
-		namespaces, err = c.listMatchingNamespaces(podMonNSSelector)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	level.Debug(c.logger).Log("msg", "filtering namespaces to select PodMonitors from", "namespaces", strings.Join(namespaces, ","), "namespace", p.Namespace, "prometheus", p.Name)
-
-	for _, ns := range namespaces {
-		err := c.pmonInfs.ListAllByNamespace(ns, podMonSelector, func(obj interface{}) {
-			k, ok := c.accessor.MetaNamespaceKey(obj)
-			if ok {
-				podMon := obj.(*monitoringv1.PodMonitor).DeepCopy()
-				if err := k8sutil.AddTypeInformationToObject(podMon); err != nil {
-					level.Error(c.logger).Log("msg", "failed to set PodMonitor type information", "namespace", ns, "err", err)
-					return
-				}
-				podMonitors[k] = podMon
-			}
-		})
-		if err != nil {
-			return nil, errors.Wrapf(err, "failed to list pod monitors in namespace %s", ns)
-		}
-	}
-
-	var rejected int
-	res := make(map[string]*monitoringv1.PodMonitor, len(podMonitors))
-	for namespaceAndName, pm := range podMonitors {
-		var err error
-
-		for i, endpoint := range pm.Spec.PodMetricsEndpoints {
-			pmKey := fmt.Sprintf("podMonitor/%s/%s/%d", pm.GetNamespace(), pm.GetName(), i)
-
-			if err = store.AddBearerToken(ctx, pm.GetNamespace(), endpoint.BearerTokenSecret, pmKey); err != nil {
-				break
-			}
-
-			if err = store.AddBasicAuth(ctx, pm.GetNamespace(), endpoint.BasicAuth, pmKey); err != nil {
-				break
-			}
-
-			if endpoint.TLSConfig != nil {
-				if err = store.AddSafeTLSConfig(ctx, pm.GetNamespace(), &endpoint.TLSConfig.SafeTLSConfig); err != nil {
-					break
-				}
-			}
-
-			if err = store.AddOAuth2(ctx, pm.GetNamespace(), endpoint.OAuth2, pmKey); err != nil {
-				break
-			}
-
-			pmAuthKey := fmt.Sprintf("podMonitor/auth/%s/%s/%d", pm.GetNamespace(), pm.GetName(), i)
-			if err = store.AddSafeAuthorizationCredentials(ctx, pm.GetNamespace(), endpoint.Authorization, pmAuthKey); err != nil {
-				break
-			}
-
-			if err = validateScrapeIntervalAndTimeout(p, endpoint.Interval, endpoint.ScrapeTimeout); err != nil {
-				break
-			}
-
-			for _, rl := range endpoint.RelabelConfigs {
-				if rl.Action != "" {
-					if err = validateRelabelConfig(*p, *rl); err != nil {
-						break
-					}
-				}
-			}
-
-			for _, rl := range endpoint.MetricRelabelConfigs {
-				if rl.Action != "" {
-					if err = validateRelabelConfig(*p, *rl); err != nil {
-						break
-					}
-				}
-			}
-		}
-
-		if err != nil {
-			rejected++
-			level.Warn(c.logger).Log(
-				"msg", "skipping podmonitor",
-				"error", err.Error(),
-				"podmonitor", namespaceAndName,
-				"namespace", p.Namespace,
-				"prometheus", p.Name,
-			)
-			continue
-		}
-
-		res[namespaceAndName] = pm
-	}
-
-	pmKeys := []string{}
-	for k := range res {
-		pmKeys = append(pmKeys, k)
-	}
-	level.Debug(c.logger).Log("msg", "selected PodMonitors", "podmonitors", strings.Join(pmKeys, ","), "namespace", p.Namespace, "prometheus", p.Name)
-
-	if pKey, ok := c.accessor.MetaNamespaceKey(p); ok {
-		c.metrics.SetSelectedResources(pKey, monitoringv1.PodMonitorsKind, len(res))
-		c.metrics.SetRejectedResources(pKey, monitoringv1.PodMonitorsKind, rejected)
-	}
-
-	return res, nil
-}
-
-func (c *Operator) selectProbes(ctx context.Context, p *monitoringv1.Prometheus, store *assets.Store) (map[string]*monitoringv1.Probe, error) {
-	namespaces := []string{}
-	// Selected object might overlap, deduplicate them by `<namespace>/<name>`.
-	probes := make(map[string]*monitoringv1.Probe)
-
-	bMonSelector, err := metav1.LabelSelectorAsSelector(p.Spec.ProbeSelector)
-	if err != nil {
-		return nil, err
-	}
-
-	// If 'ProbeNamespaceSelector' is nil only check own namespace.
-	if p.Spec.ProbeNamespaceSelector == nil {
-		namespaces = append(namespaces, p.Namespace)
-	} else {
-		bMonNSSelector, err := metav1.LabelSelectorAsSelector(p.Spec.ProbeNamespaceSelector)
-		if err != nil {
-			return nil, err
-		}
-
-		namespaces, err = c.listMatchingNamespaces(bMonNSSelector)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	level.Debug(c.logger).Log("msg", "filtering namespaces to select Probes from", "namespaces", strings.Join(namespaces, ","), "namespace", p.Namespace, "prometheus", p.Name)
-
-	for _, ns := range namespaces {
-		err := c.probeInfs.ListAllByNamespace(ns, bMonSelector, func(obj interface{}) {
-			if k, ok := c.accessor.MetaNamespaceKey(obj); ok {
-				probe := obj.(*monitoringv1.Probe).DeepCopy()
-				if err := k8sutil.AddTypeInformationToObject(probe); err != nil {
-					level.Error(c.logger).Log("msg", "failed to set Probe type information", "namespace", ns, "err", err)
-					return
-				}
-				probes[k] = probe
-			}
-		})
-		if err != nil {
-			return nil, errors.Wrapf(err, "failed to list probes in namespace %s", ns)
-		}
-	}
-
-	var rejected int
-	res := make(map[string]*monitoringv1.Probe, len(probes))
-
-	for probeName, probe := range probes {
-		rejectFn := func(probe *monitoringv1.Probe, err error) {
-			rejected++
-			level.Warn(c.logger).Log(
-				"msg", "skipping probe",
-				"error", err.Error(),
-				"probe", probe,
-				"namespace", p.Namespace,
-				"prometheus", p.Name,
-			)
-		}
-
-		if err = probe.Spec.Targets.Validate(); err != nil {
-			rejectFn(probe, err)
-			continue
-		}
-
-		pnKey := fmt.Sprintf("probe/%s/%s", probe.GetNamespace(), probe.GetName())
-		if err = store.AddBearerToken(ctx, probe.GetNamespace(), probe.Spec.BearerTokenSecret, pnKey); err != nil {
-			rejectFn(probe, err)
-			continue
-		}
-
-		if err = store.AddBasicAuth(ctx, probe.GetNamespace(), probe.Spec.BasicAuth, pnKey); err != nil {
-			rejectFn(probe, err)
-			continue
-		}
-
-		if probe.Spec.TLSConfig != nil {
-			if err = store.AddSafeTLSConfig(ctx, probe.GetNamespace(), &probe.Spec.TLSConfig.SafeTLSConfig); err != nil {
-				rejectFn(probe, err)
-				continue
-			}
-		}
-		pnAuthKey := fmt.Sprintf("probe/auth/%s/%s", probe.GetNamespace(), probe.GetName())
-		if err = store.AddSafeAuthorizationCredentials(ctx, probe.GetNamespace(), probe.Spec.Authorization, pnAuthKey); err != nil {
-			rejectFn(probe, err)
-			continue
-		}
-
-		if err = store.AddOAuth2(ctx, probe.GetNamespace(), probe.Spec.OAuth2, pnKey); err != nil {
-			rejectFn(probe, err)
-			continue
-		}
-
-		if err = validateScrapeIntervalAndTimeout(p, probe.Spec.Interval, probe.Spec.ScrapeTimeout); err != nil {
-			rejectFn(probe, err)
-			continue
-		}
-
-		for _, rl := range probe.Spec.MetricRelabelConfigs {
-			if rl.Action != "" {
-				if err = validateRelabelConfig(*p, *rl); err != nil {
-					rejectFn(probe, err)
-					continue
-				}
-			}
-		}
-		if err = validateProberURL(probe.Spec.ProberSpec.URL); err != nil {
-			err := errors.Wrapf(err, "%s url specified in proberSpec is invalid, it should be of the format `hostname` or `hostname:port`", probe.Spec.ProberSpec.URL)
-			rejectFn(probe, err)
-			continue
-		}
-		res[probeName] = probe
-	}
-
-	probeKeys := make([]string, 0)
-	for k := range res {
-		probeKeys = append(probeKeys, k)
-	}
-	level.Debug(c.logger).Log("msg", "selected Probes", "probes", strings.Join(probeKeys, ","), "namespace", p.Namespace, "prometheus", p.Name)
-
-	if pKey, ok := c.accessor.MetaNamespaceKey(p); ok {
-		c.metrics.SetSelectedResources(pKey, monitoringv1.ProbesKind, len(res))
-		c.metrics.SetRejectedResources(pKey, monitoringv1.ProbesKind, rejected)
-	}
-
-	return res, nil
-}
-
-func testForArbitraryFSAccess(e monitoringv1.Endpoint) error {
-	if e.BearerTokenFile != "" {
-		return errors.New("it accesses file system via bearer token file which Prometheus specification prohibits")
-	}
-
-	tlsConf := e.TLSConfig
-	if tlsConf == nil {
-		return nil
-	}
-
-	if tlsConf.CAFile != "" || tlsConf.CertFile != "" || tlsConf.KeyFile != "" {
-		return errors.New("it accesses file system via tls config which Prometheus specification prohibits")
-	}
-
-	return nil
-}
-
-// listMatchingNamespaces lists all the namespaces that match the provided
-// selector.
-func (c *Operator) listMatchingNamespaces(selector labels.Selector) ([]string, error) {
-	var ns []string
-	err := cache.ListAll(c.nsMonInf.GetStore(), selector, func(obj interface{}) {
-		ns = append(ns, obj.(*v1.Namespace).Name)
-	})
-	if err != nil {
-		return nil, errors.Wrap(err, "failed to list namespaces")
-	}
-	return ns, nil
-}
-
 // validateRemoteWriteSpec checks that mutually exclusive configurations are not
 // included in the Prometheus remoteWrite configuration section.
 // Reference:
@@ -2248,106 +1831,4 @@ func validateRemoteWriteSpec(spec monitoringv1.RemoteWriteSpec) error {
 	}
 
 	return nil
-}
-
-func validateRelabelConfig(p monitoringv1.Prometheus, rc monitoringv1.RelabelConfig) error {
-	relabelTarget := regexp.MustCompile(`^(?:(?:[a-zA-Z_]|\$(?:\{\w+\}|\w+))+\w*)+$`)
-	promVersion := operator.StringValOrDefault(p.Spec.Version, operator.DefaultPrometheusVersion)
-	version, err := semver.ParseTolerant(promVersion)
-	if err != nil {
-		return errors.Wrap(err, "failed to parse Prometheus version")
-	}
-	minimumVersionCaseActions := version.GTE(semver.MustParse("2.36.0"))
-	minimumVersionEqualActions := version.GTE(semver.MustParse("2.41.0"))
-
-	if (rc.Action == string(relabel.Lowercase) || rc.Action == string(relabel.Uppercase)) && !minimumVersionCaseActions {
-		return errors.Errorf("%s relabel action is only supported from Prometheus version 2.36.0", rc.Action)
-	}
-
-	if (rc.Action == string(relabel.KeepEqual) || rc.Action == string(relabel.DropEqual)) && !minimumVersionEqualActions {
-		return errors.Errorf("%s relabel action is only supported from Prometheus version 2.41.0", rc.Action)
-	}
-
-	if _, err := relabel.NewRegexp(rc.Regex); err != nil {
-		return errors.Wrapf(err, "invalid regex %s for relabel configuration", rc.Regex)
-	}
-
-	if rc.Modulus == 0 && rc.Action == string(relabel.HashMod) {
-		return errors.Errorf("relabel configuration for hashmod requires non-zero modulus")
-	}
-
-	if (rc.Action == string(relabel.Replace) || rc.Action == string(relabel.HashMod) || rc.Action == string(relabel.Lowercase) || rc.Action == string(relabel.Uppercase) || rc.Action == string(relabel.KeepEqual) || rc.Action == string(relabel.DropEqual)) && rc.TargetLabel == "" {
-		return errors.Errorf("relabel configuration for %s action needs targetLabel value", rc.Action)
-	}
-
-	if (rc.Action == string(relabel.Replace) || rc.Action == string(relabel.Lowercase) || rc.Action == string(relabel.Uppercase) || rc.Action == string(relabel.KeepEqual) || rc.Action == string(relabel.DropEqual)) && !relabelTarget.MatchString(rc.TargetLabel) {
-		return errors.Errorf("%q is invalid 'target_label' for %s action", rc.TargetLabel, rc.Action)
-	}
-
-	if (rc.Action == string(relabel.Lowercase) || rc.Action == string(relabel.Uppercase) || rc.Action == string(relabel.KeepEqual) || rc.Action == string(relabel.DropEqual)) && !(rc.Replacement == relabel.DefaultRelabelConfig.Replacement || rc.Replacement == "") {
-		return errors.Errorf("'replacement' can not be set for %s action", rc.Action)
-	}
-
-	if rc.Action == string(relabel.LabelMap) {
-		if rc.Replacement != "" && !relabelTarget.MatchString(rc.Replacement) {
-			return errors.Errorf("%q is invalid 'replacement' for %s action", rc.Replacement, rc.Action)
-		}
-	}
-
-	if rc.Action == string(relabel.HashMod) && !model.LabelName(rc.TargetLabel).IsValid() {
-		return errors.Errorf("%q is invalid 'target_label' for %s action", rc.TargetLabel, rc.Action)
-	}
-
-	if rc.Action == string(relabel.KeepEqual) || rc.Action == string(relabel.DropEqual) {
-		if !(rc.Regex == "" || rc.Regex == relabel.DefaultRelabelConfig.Regex.String()) ||
-			!(rc.Modulus == uint64(0) ||
-				rc.Modulus == relabel.DefaultRelabelConfig.Modulus) ||
-			!(rc.Separator == "" ||
-				rc.Separator == relabel.DefaultRelabelConfig.Separator) ||
-			!(rc.Replacement == relabel.DefaultRelabelConfig.Replacement ||
-				rc.Replacement == "") {
-			return errors.Errorf("%s action requires only 'source_labels' and `target_label`, and no other fields", rc.Action)
-		}
-	}
-
-	if rc.Action == string(relabel.LabelDrop) || rc.Action == string(relabel.LabelKeep) {
-		if len(rc.SourceLabels) != 0 ||
-			!(rc.TargetLabel == "" ||
-				rc.TargetLabel == relabel.DefaultRelabelConfig.TargetLabel) ||
-			!(rc.Modulus == uint64(0) ||
-				rc.Modulus == relabel.DefaultRelabelConfig.Modulus) ||
-			!(rc.Separator == "" ||
-				rc.Separator == relabel.DefaultRelabelConfig.Separator) ||
-			!(rc.Replacement == relabel.DefaultRelabelConfig.Replacement ||
-				rc.Replacement == "") {
-			return errors.Errorf("%s action requires only 'regex', and no other fields", rc.Action)
-		}
-	}
-	return nil
-}
-
-func validateProberURL(url string) error {
-	hostPort := strings.Split(url, ":")
-
-	if !govalidator.IsHost(hostPort[0]) {
-		return errors.Errorf("invalid host: %q", hostPort[0])
-	}
-
-	// handling cases with url specified as host:port
-	if len(hostPort) > 1 {
-		if !govalidator.IsPort(hostPort[1]) {
-			return errors.Errorf("invalid port: %q", hostPort[1])
-		}
-	}
-	return nil
-}
-
-func validateScrapeIntervalAndTimeout(p *monitoringv1.Prometheus, scrapeInterval, scrapeTimeout monitoringv1.Duration) error {
-	if scrapeTimeout == "" {
-		return nil
-	}
-	if scrapeInterval == "" {
-		scrapeInterval = p.Spec.ScrapeInterval
-	}
-	return prompkg.CompareScrapeTimeoutToScrapeInterval(scrapeTimeout, scrapeInterval)
 }
